split settings into appropriate components
